+++ packages/cloud-cognitive/src/components/APIKeyDownloader/APIKeyDownloader.test.js
@@ -7,7 +7,6 @@
 
 import { render, waitFor } from '@testing-library/react';
 import React from 'react';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { APIKeyDownloader } from '.';
 
 const { name } = APIKeyDownloader;