+++ packages/cloud-cognitive/src/components/APIKeyModal/APIKeyModal.test.js
@@ -7,7 +7,6 @@
 
 import { render, fireEvent } from '@testing-library/react';
 import React from 'react';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { APIKeyModal } from '.';
 
 Object.assign(navigator, {