+++ packages/cloud-cognitive/src/components/CreateModal/CreateModal.stories.js
@@ -20,7 +20,6 @@ import {
 } from 'carbon-components-react';
 
 import { pkg } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { getStorybookPrefix } from '../../../config';
 import { CreateModal } from '.';
 import mdx from './CreateModal.mdx';