+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.js
@@ -16,8 +16,8 @@ import {
   Grid,
   Column,
   Row,
-  ButtonSet,
   Button,
+  SkeletonText,
 } from 'carbon-components-react';
 import { ActionBar } from '../ActionBar/';
 import { BreadcrumbWithOverflow } from '../BreadcrumbWithOverflow';
@@ -26,6 +26,7 @@ import { ButtonSetWithOverflow } from '../ButtonSetWithOverflow';
 import { pkg } from '../../settings';
 import { ChevronUp16 } from '@carbon/icons-react';
 import {
+  deprecateProp,
   deprecatePropUsage,
   extractShapesArray,
   prepareProps,
@@ -42,9 +43,11 @@ export let PageHeader = React.forwardRef(
       background,
       breadcrumbItems,
       className,
-      collapseExpandHeaderLabel,
       collapseHeader,
-      keepBreadcrumbAndTabs,
+      collapseHeaderLabel,
+      expandHeaderLabel,
+      collapseHeaderToggleWanted,
+      preventBreadcrumbScroll,
       navigation,
       pageActions,
       pageHeaderOffset,
@@ -52,7 +55,7 @@ export let PageHeader = React.forwardRef(
       subtitle,
       tags,
       title,
-      titleIcon: TitleIcon,
+      titleIcon,
       ...rest
     },
     ref
@@ -88,6 +91,29 @@ export let PageHeader = React.forwardRef(
     ] = useState(0);
     const [actionBarColumnWidth, setActionBarColumnWidth] = useState(0);
     const [fullyCollapsed, setFullyCollapsed] = useState(false);
+
+    /**
+     * * Title shape is used to allow title to be string or shape
+     */
+    const [titleShape, setTitleShape] = useState({});
+    useEffect(() => {
+      let newShape = { ...PageHeader.defaultProps.title };
+
+      if (title?.text) {
+        // title is in shape format
+        newShape = Object.assign(newShape, { ...title });
+      } else {
+        // title is a string
+        newShape.text = title;
+      }
+
+      if (!newShape.icon && titleIcon) {
+        // if no icon use titleIcon if supplied
+        newShape.icon = titleIcon;
+      }
+
+      setTitleShape(newShape);
+    }, [title, titleIcon]);
 
     useEffect(() => {
       let newActionBarWidth = 'initial';
@@ -208,7 +234,7 @@ export let PageHeader = React.forwardRef(
       update.titleRowSpaceAbove = 0;
 
       update.headerTopValue = navigation
-        ? keepBreadcrumbAndTabs
+        ? preventBreadcrumbScroll
           ? update.navigationRowHeight +
             update.breadcrumbRowHeight -
             update.headerHeight
@@ -242,12 +268,12 @@ export let PageHeader = React.forwardRef(
       // NOTE: The buffer is used to add space between the bottom of the header and the last content
 
       // No navigation and title row not pre-collapsed
-      // and only one of tags or (subtitle or available space)
+      // and zero or one of tags or (subtitle or available space)
       setLastRowBufferActive(
-        !navigation &&
-          !preCollapseTitleRow &&
-          (title || pageActions) &&
-          !tags !== !(subtitle || availableSpace)
+        !(navigation || tags) &&
+          (((title || pageActions) && !preCollapseTitleRow) ||
+            subtitle ||
+            availableSpace)
       );
     }, [
       availableSpace,
@@ -304,7 +330,7 @@ export let PageHeader = React.forwardRef(
           [`--${blockClass}--breadcrumb-row-width-px`]: `${metrics.breadcrumbRowWidth}px`,
           [`--${blockClass}--breadcrumb-top`]: `${Math.min(
             pageHeaderOffset,
-            !keepBreadcrumbAndTabs && navigation
+            !preventBreadcrumbScroll && navigation
               ? metrics.headerHeight -
                   metrics.titleRowSpaceAbove -
                   metrics.navigationRowHeight -
@@ -316,7 +342,7 @@ export let PageHeader = React.forwardRef(
         };
       });
     }, [
-      keepBreadcrumbAndTabs,
+      preventBreadcrumbScroll,
       metrics,
       metrics.breadcrumbRowHeight,
       metrics.breadcrumbRowSpaceBelow,
@@ -332,25 +358,39 @@ export let PageHeader = React.forwardRef(
       tags,
     ]);
 
-    useEffect(() => {
-      setFullyCollapsed(
-        scrollYValue + metrics.headerTopValue + pageHeaderOffset >= 0
-      );
-    }, [metrics.headerTopValue, pageHeaderOffset, scrollYValue]);
-
     useWindowScroll(
       // on scroll or various layout changes check updates if needed
       ({ current }) => {
-        checkUpdateVerticalSpace();
+        const fullyCollapsed =
+          current.scrollY + metrics.headerTopValue + pageHeaderOffset >= 0;
+        setFullyCollapsed(fullyCollapsed);
+
+        // set offset for tagset tooltip
+        const tagsetTooltipOffset = fullyCollapsed
+          ? metrics.headerHeight + metrics.headerTopValue + pageHeaderOffset
+          : metrics.headerHeight + pageHeaderOffset;
+
+        document.documentElement.style.setProperty(
+          `--${blockClass}--tagset-tooltip-position`,
+          fullyCollapsed ? 'fixed' : 'absolute'
+        );
+
+        document.documentElement.style.setProperty(
+          `--${blockClass}--tagset-tooltip-offset`,
+          `${tagsetTooltipOffset}px`
+        );
+
         setScrollYValue(current.scrollY);
       },
       [
         actionBarItems,
         availableSpace,
         breadcrumbItems,
-        keepBreadcrumbAndTabs,
+        metrics.headerHeight,
+        metrics.headerTopValue,
         navigation,
         pageActions,
+        pageHeaderOffset,
         subtitle,
         tags,
         title,
@@ -364,7 +404,7 @@ export let PageHeader = React.forwardRef(
       actionBarItems,
       availableSpace,
       breadcrumbItems,
-      keepBreadcrumbAndTabs,
+      preventBreadcrumbScroll,
       navigation,
       pageActions,
       subtitle,
@@ -438,7 +478,7 @@ export let PageHeader = React.forwardRef(
         [`--${blockClass}--background-opacity`]: result,
       }));
       setBackgroundOpacity(result);
-      setHasCollapseButton(result > 0);
+      setHasCollapseButton(collapseHeaderToggleWanted && result > 0);
     }, [
       actionBarItems,
       background,
@@ -447,6 +487,7 @@ export let PageHeader = React.forwardRef(
       metrics.headerHeight,
       navigation,
       scrollYValue,
+      collapseHeaderToggleWanted,
       tags,
     ]);
 
@@ -491,6 +532,12 @@ export let PageHeader = React.forwardRef(
       // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [collapseHeader]);
 
+    const {
+      text: titleText,
+      icon: TitleIcon,
+      loading: titleLoading,
+    } = titleShape;
+
     return (
       <ReactResizeDetector handleHeight onResize={handleResize}>
         <section
@@ -509,11 +556,9 @@ export let PageHeader = React.forwardRef(
             {hasBreadcrumbRow ? (
               <Row
                 className={cx(`${blockClass}__breadcrumb-row`, {
-                  [`${blockClass}__breadcrumb-row--with-actions`]: hasActionBar,
                   [`${blockClass}__breadcrumb-row--next-to-tabs`]: nextToTabsCheck(),
                   [`${blockClass}__breadcrumb-row--has-breadcrumbs`]: breadcrumbItems,
-                  [`${blockClass}__breadcrumb-row--has-action-bar`]:
-                    actionBarItems || pageActions,
+                  [`${blockClass}__breadcrumb-row--has-action-bar`]: hasActionBar,
                 })}>
                 <div className={`${blockClass}__breadcrumb-row--container`}>
                   <Column
@@ -538,7 +583,7 @@ export let PageHeader = React.forwardRef(
                                 [`${blockClass}__breadcrumb-title--pre-collapsed`]: preCollapseTitleRow,
                               },
                             ])}>
-                            {title}
+                            {titleLoading ? <SkeletonText /> : titleText}
                           </BreadcrumbItem>
                         ) : (
                           ''
@@ -613,12 +658,18 @@ export let PageHeader = React.forwardRef(
                       className={cx(`${blockClass}__title`, {
                         [`${blockClass}__title--fades`]: hasBreadcrumbRow,
                       })}>
-                      {TitleIcon ? (
+                      {TitleIcon && !titleLoading ? (
                         <TitleIcon className={`${blockClass}__title-icon`} />
-                      ) : (
-                        ''
-                      )}
-                      <span title={title}>{title}</span>
+                      ) : null}
+                      <span title={!titleLoading ? titleText : null}>
+                        {titleLoading ? (
+                          <SkeletonText
+                            className={`${blockClass}__title-skeleton`}
+                          />
+                        ) : (
+                          titleText
+                        )}
+                      </span>
                     </div>
                   ) : null}
                 </Column>
@@ -628,20 +679,11 @@ export let PageHeader = React.forwardRef(
                     className={cx(`${blockClass}__page-actions`, {
                       [`${blockClass}__page-actions--in-breadcrumb`]: pageActionsInBreadcrumbRow,
                     })}>
-                    <ButtonSet
-                      className={`${blockClass}__page-actions-container`}>
-                      {pageActionsItemArray.map(
-                        ({ kind, label, onClick, ...rest }, index) => (
-                          <Button
-                            {...rest}
-                            kind={kind}
-                            onClick={onClick}
-                            key={index}>
-                            {label}
-                          </Button>
-                        )
-                      )}
-                    </ButtonSet>
+                    <ButtonSetWithOverflow
+                      className={`${blockClass}__page-actions-container`}
+                      onWidthChange={handleButtonSetWidthChange}
+                      buttons={pageActionsItemArray}
+                    />
                   </Column>
                 ) : null}
               </Row>
@@ -699,7 +741,11 @@ export let PageHeader = React.forwardRef(
                       [`${blockClass}__navigation-tags--tags-only`]:
                         navigation === undefined,
                     })}>
-                    <TagSet overflowAlign="end">{tags}</TagSet>
+                    <TagSet
+                      overflowAlign="end"
+                      overflowClassName={`${blockClass}__tagset-tooltip`}>
+                      {tags}
+                    </TagSet>
                   </Column>
                 ) : null}
               </Row>
@@ -712,7 +758,9 @@ export let PageHeader = React.forwardRef(
               })}
               data-collapse={fullyCollapsed ? 'collapsed' : 'not collapsed'}
               hasIconOnly={true}
-              iconDescription={collapseExpandHeaderLabel}
+              iconDescription={
+                fullyCollapsed ? collapseHeaderLabel : expandHeaderLabel
+              }
               kind="ghost"
               onClick={handleCollapseToggle}
               renderIcon={ChevronUp16}
@@ -782,19 +830,28 @@ PageHeader.propTypes = {
    */
   className: PropTypes.string,
   /**
-   * Label/assistive text for the collapse/expand chevron
-   */
-  collapseExpandHeaderLabel: PropTypes.string,
-  /**
    * The header can as a whole be collapsed, expanded or somewhere in between.
    * This setting controls the initial value, but also takes effect on change
+   *
+   * NOTE: The header is collapsed by setting the scroll position to hide part of the header. Collapsing has no effect if there is insufficient content to scroll.
    */
   collapseHeader: PropTypes.bool,
   /**
-   * Standard behavior scrolls the breadcrumb off to leave just tabs. This
-   * option preserves vertical space for both.
+   * Label/assistive text for the collapse/expand button
+   * Default 'Collapse'
    */
-  keepBreadcrumbAndTabs: PropTypes.bool,
+  collapseHeaderLabel: PropTypes.string,
+  /**
+   * Enable the collapse header toggle.
+   *
+   * NOTE: The header is collapsed by setting the scroll position to hide part of the header. Collapsing has no effect if there is insufficient content to scroll.
+   */
+  collapseHeaderToggleWanted: PropTypes.bool,
+  /**
+   * Label/assistive text for the collapse/expand button
+   * Default 'Expand'
+   */
+  expandHeaderLabel: PropTypes.string,
   /**
    * Content for the navigation area in the PageHeader. Should
    * be a React element that is normally a Carbon Tabs component. Optional.
@@ -835,6 +892,11 @@ PageHeader.propTypes = {
    */
   preCollapseTitleRow: PropTypes.bool,
   /**
+   * Standard behavior scrolls the breadcrumb off to leave just tabs. This
+   * option preserves vertical space for both.
+   */
+  preventBreadcrumbScroll: PropTypes.bool,
+  /**
    * A subtitle or description that provides additional context to
    * identify the current page. Optional.
    */
@@ -846,21 +908,32 @@ PageHeader.propTypes = {
   tags: PropTypes.arrayOf(PropTypes.element),
   /**
    * The title of the page.
-   * Optional.
+   * Optional string or object with the following attributes: text, icon, loading
    */
-  title: PropTypes.string,
+  title: PropTypes.oneOfType([
+    PropTypes.shape({
+      text: PropTypes.string.isRequired,
+      icon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+      loading: PropTypes.bool,
+    }),
+    PropTypes.string,
+  ]),
   /**
    * An icon to be included to the left of the title text.
    * Optional.
    */
-  titleIcon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+  titleIcon: deprecateProp(
+    PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+    'Deprecated. Use title prop shape instead.'
+  ),
 };
 
 PageHeader.defaultProps = {
-  background: false,
+  background: true,
   className: '',
-  collapseExpandHeaderLabel: 'Toggle expansion',
-  keepBreadcrumbAndTabs: false,
+  collapseHeaderLabel: 'Collapse',
+  expandHeaderLabel: 'Expand',
+  preventBreadcrumbScroll: false,
   pageHeaderOffset: 0,
   preCollapseTitleRow: false,
 };