+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.test.js
@@ -10,7 +10,6 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { pkg, carbon } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 
 import {
   BreadcrumbItem,
@@ -108,61 +107,67 @@ const tags = [
     A tag
   </Tag>,
 ];
-const title = 'Page title';
+const titleObj = { text: 'Page title', loading: false, icon: Bee32 };
+const titleString = 'Page title';
 
 import uuidv4 from '../../global/js/utils/uuidv4';
 import { prepareProps } from '../../global/js/utils/props-helper';
 jest.mock('../../global/js/utils/uuidv4');
 
 const sizes = {
-  'bx--btn': {
+  [`${carbon.prefix}--btn`]: {
     get offsetWidth() {
       return 200;
     },
   },
-  'exp--page-header': {
+  [`${blockClass}`]: {
     get offsetWidth() {
       return window.innerWidth;
     },
   },
-  'exp--page-header__breadcrumb-row': {
+  [`${blockClass}__breadcrumb-row`]: {
     get offsetWidth() {
       return window.innerWidth;
     },
   },
-  'exp--breadcrumb-with-overflow': {
+  [`${pkg.prefix}--breadcrumb-with-overflow`]: {
     get offsetWidth() {
       return window.innerWidth * 0.6;
     },
   },
-  'exp--tag-set': {
+  [`${pkg.prefix}--tag-set`]: {
     get offsetWidth() {
       return window.innerWidth * 0.25;
     },
   },
-  'exp--tag-set__sizing-tag': {
+  [`${pkg.prefix}--tag-set__sizing-tag`]: {
     get offsetWidth() {
       return window.innerWidth * 0.05;
     },
   },
-  'exp--button-set-with-overflow__button-container': {
+  [`${pkg.prefix}--button-set-with-overflow__button-container`]: {
     get offsetWidth() {
       return window.innerWidth * 0.4;
     },
   },
-  'exp--button-set-with-overflow': {
+  [`${pkg.prefix}--button-set-with-overflow`]: {
     get offsetWidth() {
       return window.innerWidth * 0.4;
     },
   },
-  'bx--breadcrumb-item': {
+  [`${carbon.prefix}--breadcrumb-item`]: {
     get offsetWidth() {
       return 200;
     },
   },
-  'exp--action-bar__displayed-items': {
+  [`${pkg.prefix}--action-bar__displayed-items`]: {
     get offsetWidth() {
       return window.innerWidth * 0.3;
+    },
+  },
+  [`${blockClass}__breadcrumb-title`]: {
+    get offsetWidth() {
+      return window.innerWidth * 0.2;
     },
   },
 };
@@ -175,7 +180,7 @@ const testSizes = (el, property) => {
       return val;
     }
   }
-  console.log(property, el.outerHTML);
+  console.log('xxx', property, el.outerHTML);
   return -1;
 };
 
@@ -189,7 +194,11 @@ const testProps = {
   pageActions,
   subtitle,
   tags,
-  title,
+  title: titleObj,
+};
+
+const testPropsAltTitle = {
+  title: titleString,
   titleIcon: Bee32,
 };
 
@@ -207,7 +216,6 @@ describe('PageHeader', () => {
           let width = testSizes(this, 'offsetWidth');
 
           if (width < 0) {
-            console.log(this.outerHTML);
             width = window.innerWidth;
           }
 
@@ -246,7 +254,7 @@ describe('PageHeader', () => {
     const header = screen.getByTestId(dataTestid);
     expect(header).toHaveClass(blockClass);
 
-    expect(header).not.toHaveClass(`${blockClass}--show-background`);
+    expect(header).toHaveClass(`${blockClass}--show-background`);
 
     expect(header).not.toHaveClass(classNames[0]);
     expect(header).not.toHaveClass(classNames[1]);
@@ -268,7 +276,7 @@ describe('PageHeader', () => {
     expect(screen.queryByText(subtitle)).toBeNull();
     expect(screen.queryAllByTestId('tags')).toHaveLength(0);
     expect(document.querySelectorAll(`.${blockClass}__title`)).toHaveLength(0);
-    expect(screen.queryByText(title)).toBeNull();
+    expect(screen.queryByText(titleObj.text)).toBeNull();
     expect(
       document.querySelectorAll(`.${blockClass}__title-icon`)
     ).toHaveLength(0);
@@ -318,7 +326,7 @@ describe('PageHeader', () => {
     ).toHaveLength(4);
     expect(document.querySelectorAll(`.${blockClass}__title`)).toHaveLength(1);
     expect(document.querySelector(`.${blockClass}__title`).textContent).toEqual(
-      title
+      titleObj.text
     );
     expect(
       document.querySelectorAll(`.${blockClass}__title-icon`)
@@ -342,7 +350,9 @@ describe('PageHeader', () => {
 
     render(<PageHeader pageActions={pageActionsDepTest} />);
 
-    screen.getByText('Primary button');
+    screen.getByText('Primary button', {
+      selector: `.${blockClass}__page-actions .${pkg.prefix}--button-set-with-overflow__button-container:not(.${pkg.prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
+    });
 
     expect(warn).toBeCalledWith(
       "The usage of prop 'pageActions' of 'PageHeader' has been changed and you should update. Expects an array of objects with the following properties: label and onClick."
@@ -356,7 +366,9 @@ describe('PageHeader', () => {
 
     render(<PageHeader pageActions={pageActionsDepTest2} />);
 
-    screen.getByText('Primary button');
+    screen.getByText('Primary button', {
+      selector: `.${blockClass}__page-actions .${pkg.prefix}--button-set-with-overflow__button-container:not(.${pkg.prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
+    });
 
     expect(warn).toBeCalledWith(
       "The usage of prop 'pageActions' of 'PageHeader' has been changed and you should update. Expects an array of objects with the following properties: label and onClick."
@@ -379,14 +391,18 @@ describe('PageHeader', () => {
 
   test('collapse button works', () => {
     const dataTestid = uuidv4();
-    render(<PageHeader {...testProps} data-testid={dataTestid} />);
+    render(
+      <PageHeader
+        {...testProps}
+        collapseHeaderLabel="Toggle collapse"
+        expandHeaderLabel="Toggle expand"
+        collapseHeaderToggleWanted={true}
+        data-testid={dataTestid}
+      />
+    );
 
-    // console.dir(screen.getByRole('region')); // section should be a region https://fae.disability.illinois.edu/rulesets/ROLE_5/
-    // const header = screen.getByTestId(dataTestid);
-    const collapseButton = screen.getByText('Toggle expansion');
+    const collapseButton = screen.getByText('Toggle collapse');
 
-    // const prevCalls = window.scrollTo.callCount;
-    // console.dir(window.scrollTo);
     window.scrollTo.mockReset();
     expect(window.scrollTo).not.toHaveBeenCalled();
     userEvent.click(collapseButton);
@@ -455,5 +471,21 @@ describe('PageHeader', () => {
       `.${blockClass}__action-bar`
     );
     expect(actionBarItems).toHaveLength(1);
+  });
+
+  test('renders  title when using separate string and icon', () => {
+    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
+
+    render(<PageHeader {...testPropsAltTitle} />);
+
+    expect(warn).toBeCalledWith(
+      "The prop 'titleIcon' of 'PageHeader' has been deprecated and will soon be removed. Deprecated. Use title prop shape instead."
+    );
+
+    screen.getByText(titleString, { selector: `.${blockClass}__title span` });
+
+    expect(
+      document.querySelectorAll(`.${blockClass}__title-icon`)
+    ).toHaveLength(1);
   });
 });